Feature/dto templete

+++ src/ValueObject/OrderForm.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Coderun\BuyOneClick\ValueObject;
 
-use Coderun\BuyOneClick\LoadFile;
 use Coderun\BuyOneClick\Options\Notification as NotificationOptions;
 use Coderun\BuyOneClick\Repository\Order;
+use Coderun\BuyOneClick\SimpleDataObjects\DownloadedFile;
 use Coderun\BuyOneClick\Utils\Uuid as UuidUtils;
 use Coderun\BuyOneClick\VariationsAddition;
 use WC_Data_Exception;
@@ -86,6 +86,14 @@ class OrderForm
      * @var string
      */
     protected string $orderUuid = '';
+    
+    /**
+     * Является товаром WooCommerce
+     * Товар может быть с произвольным ИД, не являясь товаром WooCommerce
+     *
+     * @var bool
+     */
+    protected bool $isWooCommerceProduct = false;
 
     /**
      * @param array<string, mixed> $formData
@@ -95,7 +103,8 @@ class OrderForm
     public function __construct(
         array $formData,
         NotificationOptions $notificationOptions,
-        bool $variationEnable = false
+        bool $variationEnable = false,
+        array $files = []
     ) {
         $this->formData = $formData;
         $this->userName = $this->formDateParse('txtname');
@@ -104,7 +113,8 @@ class OrderForm
         $this->userComment = $this->formDateParse('message');
         $this->orderComment = $this->formDateParse('message');
         $this->productId = (int)$this->formDateParse('idtovar');
-        $this->productUrl = get_the_permalink($this->productId);
+        $this->isWooCommerceProduct = boolval(wc_get_product($this->productId));
+        $this->productUrl = strval(get_the_permalink($this->productId));
         $this->productName = $this->formDateParse('nametovar');
         $this->productOriginalName = $this->formDateParse('nametovar');
         $this->productPrice = (float)$this->formDateParse('pricetovar');
@@ -116,15 +126,14 @@ class OrderForm
         $this->formsField = $this->formDateLegacyParse();
         $this->orderTime = current_time('mysql');
         $this->custom = (int)$this->formDateParse('custom');
-        $this->files = array_map('array_filter', $_FILES['files'] ?? []);
+        $this->files = $files;
         $this->quantityProduct = $this->formDateParse('quantity_product') == ''
             ? 1 : intval($this->formDateParse('quantity_product'));
         $this->fillInPriceWithTax();
         if ($variationEnable) {
             $this->fillingWithVariations();
         }
-        $dataAboutUploadedFiles = LoadFile::getInstance()->load();
-        $this->filesUrlCollection = $this->collectUrlToUploadedFiles($dataAboutUploadedFiles);
+        $this->filesUrlCollection = $this->collectUrlToUploadedFiles($files);
         foreach ($this->filesUrlCollection as $fileUrl) {
             $this->filesLink = sprintf('</br> %s', $this->collectLinkToProductForUser($fileUrl));
         }
@@ -140,6 +149,9 @@ class OrderForm
      */
     private function fillInPriceWithTax(): void
     {
+        if (!$this->isWooCommerceProduct) {
+            return;
+        }
         $wcOrder = Order::getInstance()->createWooCommerceOrderWithoutSaving($this->productId);
         $this->productPriceWithTax = (float)Order::getInstance()->calculate_order_totals($wcOrder);
         $wcOrder->delete();
@@ -155,7 +167,8 @@ class OrderForm
     {
         $pluginVariations = VariationsAddition::getInstance();
         $this->variationData = $pluginVariations->getVariableProductInfo($this->getFormsField());
-        if (($variation_id = $pluginVariations->getVariationId($this->getFormsField())) > 0) {
+        $variation_id = $pluginVariations->getVariationId($this->getFormsField());
+        if ($variation_id > 0) {
             $this->productIsVariable = true;
             $this->productId = (int)$variation_id;
         }
@@ -235,7 +248,7 @@ class OrderForm
 
 
     /**
-     * @param array $files
+     * @param array<int, DownloadedFile> $files
      *
      * @return array<string,mixed>
      */
@@ -243,34 +256,9 @@ class OrderForm
     {
         $result = [];
         foreach ($files as $file) {
-            $result[] = $file['url'] ?? null;
+            $result[] = $file->url;
         }
         return array_filter($result);
-    }
-
-    /**
-     * Файлы в виде ссылок и строки
-     * @param array $files
-     * @return string
-     */
-    private function collectLinksToUploadedFiles(array $files): string
-    {
-        $result = '';
-        $count = 1;
-        foreach ($this->collectUrlToUploadedFiles($files) as $url) {
-            $url = trim($url);
-            if (strlen($url) == 0) {
-                continue;
-            }
-            $result .=sprintf(
-                '<br><a href="%s">%s %s</a>',
-                $url,
-                __('File', 'coderun-oneclickwoo'),
-                $count++
-            );
-        }
-
-        return $result;
     }
 
     /**
@@ -388,7 +376,7 @@ class OrderForm
     }
 
     /**
-     * @return array|string
+     * @return string
      */
     public function getProductName()
     {
@@ -636,9 +624,9 @@ class OrderForm
     }
 
     /**
-     * @return array|null
+     * @return array
      */
-    public function getFiles(): ?array
+    public function getFiles(): array
     {
         return $this->files;
     }
@@ -648,7 +636,7 @@ class OrderForm
      *
      * @return OrderForm
      */
-    public function setFiles(?array $files): OrderForm
+    public function setFiles(array $files): OrderForm
     {
         $this->files = $files;
         return $this;
@@ -786,4 +774,25 @@ class OrderForm
         $this->orderUuid = $orderUuid;
         return $this;
     }
+    
+    /**
+     * @return bool
+     */
+    public function isWooCommerceProduct(): bool
+    {
+        return $this->isWooCommerceProduct;
+    }
+    
+    /**
+     * @param bool $isWooCommerceProduct
+     *
+     * @return OrderForm
+     */
+    public function setIsWooCommerceProduct(bool $isWooCommerceProduct
+    ): OrderForm {
+        $this->isWooCommerceProduct = $isWooCommerceProduct;
+        return $this;
+    }
+    
+    
 }
